Restructuring architecture
