+++ .gitignore
@@ -1,0 +1,5 @@
+/.idea/
+/.idea/*
+/.DS_Store
+/Debuggez-app-Java.iml
+/out/
