--- Project02Eclipse/README.md
@@ -1,1 +1,0 @@
-Sample project for OpenClassroom's fundamentals of Java
