+++ README.md
@@ -1,1 +1,2 @@
 # Project_DA_Java_EN_Come_to_the_Rescue_of_a_Java_Application
+## Sample project for OpenClassroom's fundamentals of Java
