+++ src/main/java/com/hemebiotech/AnalyticsCounter.java
@@ -1,4 +1,4 @@
-package com.hemebiotech.analytics;
+package com.hemebiotech;
 
 import java.io.BufferedReader;
 import java.io.FileReader;
