+++ src/main/java/com/hemebiotech/ISymptomReader.java
@@ -1,4 +1,4 @@
-package com.hemebiotech.analytics;
+package com.hemebiotech;
 
 import java.util.List;
